# ecommerce/Postman_OpenAPI_collection.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Product Controller","description":"Product Controller"},{"name":"Order controller","description":"Order controller"}],"paths":{"/api/v1/orders":{"post":{"tags":["Order controller"],"summary":"Place an order","operationId":"placeOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderRequestDto"}}},"required":true},"responses":{"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponseDto"}}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponseDto"}}}},"200":{"description":"Order placed successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponseDto"}}}}}}},"/api/v1/orders/cancel_order":{"post":{"tags":["Order controller"],"summary":"Cancel an order","operationId":"cancelOrder","parameters":[{"name":"orderReferenceId","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad request"},"404":{"description":"Not found"},"200":{"description":"Order cancelled successfully"}}}},"/api/v1/products/product_enquiry":{"get":{"tags":["Product Controller"],"summary":"Check product availability","operationId":"productEnquiry","parameters":[{"name":"productId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductEnquiryDto"}}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductEnquiryDto"}}}},"200":{"description":"Product enquiry successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductEnquiryDto"}}}}}}},"/api/v1/products/item_menu":{"get":{"tags":["Product Controller"],"summary":"Get All menu items","operationId":"getAllMenuItems","responses":{"400":{"description":"Bad request","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductDto"}}}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductDto"}}}}},"200":{"description":"Product enquiry successful","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductDto"}}}}}}}},"/api/v1/orders/check_order_status":{"get":{"tags":["Order controller"],"summary":"Check order status","operationId":"checkOrderStatus","parameters":[{"name":"orderReferenceId","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderStatusDto"}}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderStatusDto"}}}},"200":{"description":"Order status checked successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderStatusDto"}}}}}}}},"components":{"schemas":{"OrderItemDto":{"required":["productId","qty","totalPrice","uen"],"type":"object","properties":{"productId":{"type":"integer","format":"int64"},"productName":{"type":"string"},"uen":{"type":"string"},"qty":{"type":"integer","format":"int32"},"totalPrice":{"type":"number","format":"double"}}},"OrderRequestDto":{"required":["orders"],"type":"object","properties":{"orders":{"type":"array","items":{"$ref":"#/components/schemas/OrderItemDto"}}}},"OrderResponseDto":{"type":"object","properties":{"orderReferenceId":{"type":"string"}}},"ProductEnquiryDto":{"required":["productId","qty"],"type":"object","properties":{"productId":{"type":"integer","format":"int64"},"qty":{"type":"integer","format":"int32"}}},"ProductDto":{"type":"object","properties":{"product_id":{"type":"integer","format":"int64"},"shop_name":{"type":"string"},"name":{"type":"string"},"quantity":{"type":"integer","format":"int32"},"price":{"type":"number","format":"double"},"uen":{"type":"string"}}},"OrderStatusDto":{"type":"object","properties":{"orderReferenceId":{"type":"string"},"items":{"type":"array","items":{"$ref":"#/components/schemas/OrderItemDto"}},"status":{"type":"string","enum":["PENDING","PLACED","SHIPPED","DELIVERED","CANCELLED"]}}}}}}