# ecommerce/src/main/java/com/ecommerce/dto/OrderRequestDto.java
package com.ecommerce.dto;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import javax.validation.constraints.NotNull;
import java.util.List;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class OrderRequestDto {
    @NotNull(message = "order items cannot be empty")
    private List<OrderItemDto> orders;
}