# ecommerce/src/main/java/com/ecommerce/dto/ProductEnquiryDto.java
package com.ecommerce.dto;

import lombok.*;

import javax.validation.constraints.NotNull;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class ProductEnquiryDto {
    @NotNull(message = "Product ID cannot be null")
    private Long productId;
    @NotNull(message = "Quantity cannot be null")
    private Integer qty;
}